requirements.txt

mcp>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5
email-validator>=2.0
python-dotenv>=1.0.0
